moved constants to stm32u545
