+++ chips/stm32u545/src/lib.rs
@@ -8,8 +8,7 @@ pub use stm32u5xx::{adc, chip, dma, exti, gpio, pwr, rcc, tim, usart};
 
 use cortexm33::{CortexM33, CortexMVariant};
 
-pub type Trng<'a> =
-    stm32u5xx::entropy::Trng<'a, RNG_CR_CONFIG_U545, RNG_HTCR_CONFIG_U545, RNG_NSCR_CONFIG_U545>;
+pub mod entropy;
 
 #[cfg_attr(all(target_arch = "arm", target_os = "none"), used)]
 #[cfg_attr(all(target_arch = "arm", target_os = "none"), link_section = ".irqs")]
